code/LingoBuddy/pip-requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests
deep_translator
PyPDF2
chromadb
python-dotenv
sentence-transformers
yt-dlp
moviepy==1.0.3
pydub
gtts
speechrecognition
boto3
